+++ tests/image_methods_test.py
@@ -58,6 +58,16 @@ def test_adaptive_threshold():
         assert was != img.signature
 
 
+def test_annotate():
+    from wand.drawing import Drawing
+    with Image(filename='rose:') as img:
+        was = img.signature
+        with Drawing() as ctx:
+            ctx.font_size = 32
+            img.annotate('Hello', ctx, left=10, baseline=img.height-10)
+        assert was != img.signature
+
+
 def test_auto_gamma():
     with Image(filename='rose:') as img:
         was = img.signature