+++ docs/changes.rst
@@ -15,6 +15,7 @@ Version 0.5.6
 Unreleased.
 
  - Added :meth:`Image.percent_escape() <wand.image.BaseImage.percent_escape>` helper method. [:issue:`421`]
+ - Added :meth:`Image.annotate() <wand.image.BaseImage.annotate>` method. [:issue:`#418`]
 
 
 .. _changelog-0.5.5: